[REFACTOR/#193] 코리 반영 수정
