+++ core-ui/src/main/java/com/going/ui/extension/IntExt.kt
@@ -1,9 +1,21 @@
 package com.going.ui.extension
 
 import android.content.Context
+import android.content.Intent
+import android.os.Build
 import android.util.TypedValue
 
 fun Int.dpToPx(context: Context): Int {
     val metrics = context.resources.displayMetrics
     return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this.toFloat(), metrics).toInt()
 }
+
+fun <T> Intent.getParcelable(name: String, clazz: Class<T>): T? =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        getParcelableExtra(name, clazz)
+    } else {
+        getParcelableExtra(name)
+    }
+
+
+
