--- core-ui/src/main/java/com/going/ui/extension/getParceable.kt
@@ -1,11 +1,0 @@
-package com.going.ui.extension
-
-import android.content.Intent
-import android.os.Build
-
-fun <T> Intent.getParcelable(name: String, clazz: Class<T>): T? =
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-        getParcelableExtra(name, clazz)
-    } else {
-        getParcelableExtra(name)
-    }
